+++ main.tf
@@ -1,4 +1,5 @@
 data "aws_caller_identity" "current" {}
+data "aws_partition" "current" {}
 
 data "aws_subnet" "runners" {
   id = length(var.subnet_id) > 0 ? var.subnet_id : var.subnet_id_runners
@@ -40,7 +41,10 @@ locals {
       logging             = var.enable_cloudwatch_logging ? local.logging_user_data : ""
       gitlab_runner       = local.template_gitlab_runner
       user_data_trace_log = var.enable_runner_user_data_trace_log
+      yum_update          = var.runner_yum_update ? local.file_yum_update : ""
   })
+
+  file_yum_update = file("${path.module}/template/yum_update.tpl")
 
   template_eip = templatefile("${path.module}/template/eip.tpl", {
     eip = join(",", aws_eip.gitlab_runner.*.public_ip)
@@ -73,32 +77,25 @@ locals {
 
   template_runner_config = templatefile("${path.module}/template/runner-config.tpl",
     {
-      aws_region                  = var.aws_region
-      gitlab_url                  = var.runners_gitlab_url
-      runners_vpc_id              = var.vpc_id
-      runners_subnet_id           = length(var.subnet_id) > 0 ? var.subnet_id : var.subnet_id_runners
-      runners_aws_zone            = data.aws_availability_zone.runners.name_suffix
-      runners_instance_type       = var.docker_machine_instance_type
-      runners_spot_price_bid      = var.docker_machine_spot_price_bid == "on-demand-price" ? "" : var.docker_machine_spot_price_bid
-      runners_ami                 = data.aws_ami.docker-machine.id
-      runners_security_group_name = aws_security_group.docker_machine.name
-      runners_monitoring          = var.runners_monitoring
-      runners_ebs_optimized       = var.runners_ebs_optimized
-      runners_instance_profile    = aws_iam_instance_profile.docker_machine.name
-      runners_additional_volumes  = local.runners_additional_volumes
-      docker_machine_options      = length(local.docker_machine_options_string) == 1 ? "" : local.docker_machine_options_string
-      runners_name                = var.runners_name
-      runners_tags = replace(replace(var.overrides["name_docker_machine_runners"] == "" ? format(
-        "Name,%s-docker-machine,%s,%s",
-        var.environment,
-        local.tags_string,
-        local.runner_tags_string,
-        ) : format(
-        "%s,%s,Name,%s",
-        local.tags_string,
-        local.runner_tags_string,
-        var.overrides["name_docker_machine_runners"],
-      ), ",,", ","), "/,$/", "")
+      aws_region                        = var.aws_region
+      gitlab_url                        = var.runners_gitlab_url
+      gitlab_clone_url                  = var.runners_clone_url
+      runners_extra_hosts               = var.runners_extra_hosts
+      runners_vpc_id                    = var.vpc_id
+      runners_subnet_id                 = length(var.subnet_id) > 0 ? var.subnet_id : var.subnet_id_runners
+      runners_aws_zone                  = data.aws_availability_zone.runners.name_suffix
+      runners_instance_type             = var.docker_machine_instance_type
+      runners_spot_price_bid            = var.docker_machine_spot_price_bid == "on-demand-price" ? "" : var.docker_machine_spot_price_bid
+      runners_ami                       = data.aws_ami.docker_machine.id
+      runners_security_group_name       = aws_security_group.docker_machine.name
+      runners_monitoring                = var.runners_monitoring
+      runners_ebs_optimized             = var.runners_ebs_optimized
+      runners_instance_profile          = aws_iam_instance_profile.docker_machine.name
+      runners_additional_volumes        = local.runners_additional_volumes
+      docker_machine_options            = length(local.docker_machine_options_string) == 1 ? "" : local.docker_machine_options_string
+      docker_machine_name               = format("%s-%s", local.runner_tags_merged["Name"], "%s") # %s is always needed
+      runners_name                      = var.runners_name
+      runners_tags                      = replace(replace(local.runner_tags_string, ",,", ","), "/,$/", "")
       runners_token                     = var.runners_token
       runners_executor                  = var.runners_executor
       runners_limit                     = var.runners_limit
@@ -109,12 +106,13 @@ locals {
       runners_docker_runtime            = var.runners_docker_runtime
       runners_helper_image              = var.runners_helper_image
       runners_shm_size                  = var.runners_shm_size
-      runners_pull_policy               = var.runners_pull_policy
+      runners_pull_policies             = local.runners_pull_policies
       runners_idle_count                = var.runners_idle_count
       runners_idle_time                 = var.runners_idle_time
       runners_max_builds                = local.runners_max_builds_string
       runners_machine_autoscaling       = local.runners_machine_autoscaling
       runners_root_size                 = var.runners_root_size
+      runners_volume_type               = var.runners_volume_type
       runners_iam_instance_profile_name = var.runners_iam_instance_profile_name
       runners_use_private_address_only  = var.runners_use_private_address
       runners_use_private_address       = !var.runners_use_private_address
@@ -128,6 +126,7 @@ locals {
       runners_check_interval            = var.runners_check_interval
       runners_volumes_tmpfs             = join("\n", [for v in var.runners_volumes_tmpfs : format("\"%s\" = \"%s\"", v.volume, v.options)])
       runners_services_volumes_tmpfs    = join("\n", [for v in var.runners_services_volumes_tmpfs : format("\"%s\" = \"%s\"", v.volume, v.options)])
+      runners_docker_services           = local.runners_docker_services
       bucket_name                       = local.bucket_name
       shared_cache                      = var.cache_shared
       sentry_dsn                        = var.sentry_dsn
@@ -137,7 +136,7 @@ locals {
   )
 }
 
-data "aws_ami" "docker-machine" {
+data "aws_ami" "docker_machine" {
   most_recent = "true"
 
   dynamic "filter" {
@@ -187,6 +186,9 @@ resource "aws_autoscaling_group" "gitlab_runner_instance" {
   timeouts {
     delete = var.asg_delete_timeout
   }
+  lifecycle {
+    ignore_changes = [min_size, max_size, desired_capacity]
+  }
 }
 
 resource "aws_autoscaling_schedule" "scale_in" {
@@ -194,9 +196,9 @@ resource "aws_autoscaling_schedule" "scale_in" {
   autoscaling_group_name = aws_autoscaling_group.gitlab_runner_instance.name
   scheduled_action_name  = "scale_in-${aws_autoscaling_group.gitlab_runner_instance.name}"
   recurrence             = var.schedule_config["scale_in_recurrence"]
-  min_size               = var.schedule_config["scale_in_count"]
-  desired_capacity       = var.schedule_config["scale_in_count"]
-  max_size               = var.schedule_config["scale_in_count"]
+  min_size               = try(var.schedule_config["scale_in_min_size"], var.schedule_config["scale_in_count"])
+  desired_capacity       = try(var.schedule_config["scale_in_desired_capacity"], var.schedule_config["scale_in_count"])
+  max_size               = try(var.schedule_config["scale_in_max_size"], var.schedule_config["scale_in_count"])
 }
 
 resource "aws_autoscaling_schedule" "scale_out" {
@@ -204,9 +206,9 @@ resource "aws_autoscaling_schedule" "scale_out" {
   autoscaling_group_name = aws_autoscaling_group.gitlab_runner_instance.name
   scheduled_action_name  = "scale_out-${aws_autoscaling_group.gitlab_runner_instance.name}"
   recurrence             = var.schedule_config["scale_out_recurrence"]
-  min_size               = var.schedule_config["scale_out_count"]
-  desired_capacity       = var.schedule_config["scale_out_count"]
-  max_size               = var.schedule_config["scale_out_count"]
+  min_size               = try(var.schedule_config["scale_out_min_size"], var.schedule_config["scale_out_count"])
+  desired_capacity       = try(var.schedule_config["scale_out_desired_capacity"], var.schedule_config["scale_out_count"])
+  max_size               = try(var.schedule_config["scale_out_max_size"], var.schedule_config["scale_out_count"])
 }
 
 data "aws_ami" "runner" {
@@ -226,7 +228,7 @@ data "aws_ami" "runner" {
 resource "aws_launch_template" "gitlab_runner_instance" {
   name_prefix            = local.name_runner_agent_instance
   image_id               = data.aws_ami.runner.id
-  user_data              = base64encode(local.template_user_data)
+  user_data              = base64gzip(local.template_user_data)
   instance_type          = var.instance_type
   update_default_version = true
   ebs_optimized          = var.runner_instance_ebs_optimized
@@ -246,7 +248,7 @@ resource "aws_launch_template" "gitlab_runner_instance" {
     }
   }
   iam_instance_profile {
-    name = aws_iam_instance_profile.instance.name
+    name = local.aws_iam_role_instance_name
   }
   dynamic "block_device_mappings" {
     for_each = [var.runner_root_block_device]
@@ -301,22 +303,26 @@ resource "aws_launch_template" "gitlab_runner_instance" {
 ### Create cache bucket
 ################################################################################
 locals {
-  bucket_name   = var.cache_bucket["create"] ? module.cache.bucket : lookup(var.cache_bucket, "bucket", "")
-  bucket_policy = var.cache_bucket["create"] ? module.cache.policy_arn : lookup(var.cache_bucket, "policy", "")
+  bucket_name   = var.cache_bucket["create"] ? module.cache[0].bucket : var.cache_bucket["bucket"]
+  bucket_policy = var.cache_bucket["create"] ? module.cache[0].policy_arn : var.cache_bucket["policy"]
 }
 
 module "cache" {
+  count  = var.cache_bucket["create"] ? 1 : 0
   source = "./modules/cache"
 
   environment = var.environment
   tags        = local.tags
 
-  create_cache_bucket                  = var.cache_bucket["create"]
   cache_bucket_prefix                  = var.cache_bucket_prefix
   cache_bucket_name_include_account_id = var.cache_bucket_name_include_account_id
   cache_bucket_set_random_suffix       = var.cache_bucket_set_random_suffix
   cache_bucket_versioning              = var.cache_bucket_versioning
   cache_expiration_days                = var.cache_expiration_days
+  cache_logging_bucket                 = var.cache_logging_bucket
+  cache_logging_bucket_prefix          = var.cache_logging_bucket_prefix
+
+  kms_key_id = local.kms_key
 
   name_iam_objects = local.name_iam_objects
 }
@@ -325,15 +331,18 @@ module "cache" {
 ### Trust policy
 ################################################################################
 resource "aws_iam_instance_profile" "instance" {
-  name = "${local.name_iam_objects}-instance"
-  role = aws_iam_role.instance.name
+  count = var.create_runner_iam_role ? 1 : 0
+
+  name = local.aws_iam_role_instance_name
+  role = local.aws_iam_role_instance_name
   tags = local.tags
 }
 
 resource "aws_iam_role" "instance" {
-  name                 = "${local.name_iam_objects}-instance"
+  count                = var.create_runner_iam_role ? 1 : 0
+  name                 = local.aws_iam_role_instance_name
   assume_role_policy   = length(var.instance_role_json) > 0 ? var.instance_role_json : templatefile("${path.module}/policies/instance-role-trust-policy.json", {})
-  permissions_boundary = var.permissions_boundary == "" ? null : "${var.arn_format}:iam::${data.aws_caller_identity.current.account_id}:policy/${var.permissions_boundary}"
+  permissions_boundary = var.permissions_boundary == "" ? null : "arn:${data.aws_partition.current.partition}:iam::${data.aws_caller_identity.current.account_id}:policy/${var.permissions_boundary}"
   tags                 = merge(local.tags, var.role_tags)
 }
 
@@ -343,6 +352,7 @@ resource "aws_iam_role" "instance" {
 ### iam:PassRole To pass the role from the agent to the docker machine runners
 ################################################################################
 resource "aws_iam_policy" "instance_docker_machine_policy" {
+  count       = var.create_runner_iam_role ? 1 : 0
   name        = "${local.name_iam_objects}-docker-machine"
   path        = "/"
   description = "Policy for docker machine."
@@ -354,8 +364,10 @@ resource "aws_iam_policy" "instance_docker_machine_policy" {
 }
 
 resource "aws_iam_role_policy_attachment" "instance_docker_machine_policy" {
-  role       = aws_iam_role.instance.name
-  policy_arn = aws_iam_policy.instance_docker_machine_policy.arn
+  count = var.create_runner_iam_role ? 1 : 0
+
+  role       = local.aws_iam_role_instance_name
+  policy_arn = aws_iam_policy.instance_docker_machine_policy[count.index].arn
 }
 
 ################################################################################
@@ -374,23 +386,24 @@ resource "aws_iam_policy" "instance_session_manager_policy" {
 resource "aws_iam_role_policy_attachment" "instance_session_manager_policy" {
   count = var.enable_runner_ssm_access ? 1 : 0
 
-  role       = aws_iam_role.instance.name
+  role       = local.aws_iam_role_instance_name
   policy_arn = aws_iam_policy.instance_session_manager_policy[0].arn
 }
 
 resource "aws_iam_role_policy_attachment" "instance_session_manager_aws_managed" {
   count = var.enable_runner_ssm_access ? 1 : 0
 
-  role       = aws_iam_role.instance.name
-  policy_arn = "${var.arn_format}:iam::aws:policy/AmazonSSMManagedInstanceCore"
+  role       = local.aws_iam_role_instance_name
+  policy_arn = "arn:${data.aws_partition.current.partition}:iam::aws:policy/AmazonSSMManagedInstanceCore"
 }
 
 ################################################################################
 ### Add user defined policies
 ################################################################################
 resource "aws_iam_role_policy_attachment" "user_defined_policies" {
-  count      = length(var.runner_iam_policy_arns)
-  role       = aws_iam_role.instance.name
+  count = length(var.runner_iam_policy_arns)
+
+  role       = local.aws_iam_role_instance_name
   policy_arn = var.runner_iam_policy_arns[count.index]
 }
 
@@ -398,8 +411,12 @@ resource "aws_iam_role_policy_attachment" "user_defined_policies" {
 ### Policy for the docker machine instance to access cache
 ################################################################################
 resource "aws_iam_role_policy_attachment" "docker_machine_cache_instance" {
-  count      = var.cache_bucket["create"] || lookup(var.cache_bucket, "policy", "") != "" ? 1 : 0
-  role       = aws_iam_role.instance.name
+  /* If the S3 cache adapter is configured to use an IAM instance profile, the
+     adapter uses the profile attached to the GitLab Runner machine. So do not
+     use aws_iam_role.docker_machine.name here! See https://docs.gitlab.com/runner/configuration/advanced-configuration.html */
+  count = var.cache_bucket["create"] || length(lookup(var.cache_bucket, "policy", "")) > 0 ? 1 : 0
+
+  role       = local.aws_iam_role_instance_name
   policy_arn = local.bucket_policy
 }
 
@@ -409,7 +426,7 @@ resource "aws_iam_role_policy_attachment" "docker_machine_cache_instance" {
 resource "aws_iam_role" "docker_machine" {
   name                 = "${local.name_iam_objects}-docker-machine"
   assume_role_policy   = length(var.docker_machine_role_json) > 0 ? var.docker_machine_role_json : templatefile("${path.module}/policies/instance-role-trust-policy.json", {})
-  permissions_boundary = var.permissions_boundary == "" ? null : "${var.arn_format}:iam::${data.aws_caller_identity.current.account_id}:policy/${var.permissions_boundary}"
+  permissions_boundary = var.permissions_boundary == "" ? null : "arn:${data.aws_partition.current.partition}:iam::${data.aws_caller_identity.current.account_id}:policy/${var.permissions_boundary}"
   tags                 = local.tags
 }
 
@@ -433,7 +450,7 @@ resource "aws_iam_role_policy_attachment" "docker_machine_session_manager_aws_ma
   count = var.enable_docker_machine_ssm_access ? 1 : 0
 
   role       = aws_iam_role.docker_machine.name
-  policy_arn = "${var.arn_format}:iam::aws:policy/AmazonSSMManagedInstanceCore"
+  policy_arn = "arn:${data.aws_partition.current.partition}:iam::aws:policy/AmazonSSMManagedInstanceCore"
 }
 
 ################################################################################
@@ -445,14 +462,14 @@ resource "aws_iam_policy" "service_linked_role" {
   name        = "${local.name_iam_objects}-service_linked_role"
   path        = "/"
   description = "Policy for creation of service linked roles."
-  policy      = templatefile("${path.module}/policies/service-linked-role-create-policy.json", { arn_format = var.arn_format })
+  policy      = templatefile("${path.module}/policies/service-linked-role-create-policy.json", { partition = data.aws_partition.current.partition })
   tags        = local.tags
 }
 
 resource "aws_iam_role_policy_attachment" "service_linked_role" {
   count = var.allow_iam_service_linked_role_creation ? 1 : 0
 
-  role       = aws_iam_role.instance.name
+  role       = local.aws_iam_role_instance_name
   policy_arn = aws_iam_policy.service_linked_role[0].arn
 }
 
@@ -467,13 +484,13 @@ resource "aws_iam_policy" "ssm" {
   name        = "${local.name_iam_objects}-ssm"
   path        = "/"
   description = "Policy for runner token param access via SSM"
-  policy      = templatefile("${path.module}/policies/instance-secure-parameter-role-policy.json", { arn_format = var.arn_format })
+  policy      = templatefile("${path.module}/policies/instance-secure-parameter-role-policy.json", { partition = data.aws_partition.current.partition })
   tags        = local.tags
 }
 
 resource "aws_iam_role_policy_attachment" "ssm" {
-  role       = aws_iam_role.instance.name
-  policy_arn = aws_iam_policy.ssm.arn
+  role       = local.aws_iam_role_instance_name
+  policy_arn = aws_iam_policy.ssm[0].arn
 }
 
 ################################################################################
@@ -492,7 +509,7 @@ resource "aws_iam_policy" "eip" {
 resource "aws_iam_role_policy_attachment" "eip" {
   count = var.enable_eip ? 1 : 0
 
-  role       = aws_iam_role.instance.name
+  role       = local.aws_iam_role_instance_name
   policy_arn = aws_iam_policy.eip[0].arn
 }
 
@@ -513,7 +530,9 @@ module "terminate_instances_lifecycle_function" {
   lambda_runtime                       = var.asg_terminate_lifecycle_lambda_runtime
   lifecycle_heartbeat_timeout          = var.asg_terminate_lifecycle_hook_heartbeat_timeout
   name_iam_objects                     = local.name_iam_objects
-  role_permissions_boundary            = var.permissions_boundary == "" ? null : "${var.arn_format}:iam::${data.aws_caller_identity.current.account_id}:policy/${var.permissions_boundary}"
+  role_permissions_boundary            = var.permissions_boundary == "" ? null : "arn:${data.aws_partition.current.partition}:iam::${data.aws_caller_identity.current.account_id}:policy/${var.permissions_boundary}"
   lambda_timeout                       = var.asg_terminate_lifecycle_lambda_timeout
+  kms_key_id                           = local.kms_key
+  arn_format                           = var.arn_format
   tags                                 = local.tags
 }