+++ outputs.tf
@@ -5,22 +5,22 @@ output "runner_as_group_name" {
 
 output "runner_cache_bucket_arn" {
   description = "ARN of the S3 for the build cache."
-  value       = module.cache.arn
+  value       = element(concat(module.cache.*.arn, [""]), 0)
 }
 
 output "runner_cache_bucket_name" {
   description = "Name of the S3 for the build cache."
-  value       = module.cache.bucket
+  value       = element(concat(module.cache.*.bucket, [""]), 0)
 }
 
 output "runner_agent_role_arn" {
   description = "ARN of the role used for the ec2 instance for the GitLab runner agent."
-  value       = aws_iam_role.instance.arn
+  value       = local.aws_iam_role_instance_arn
 }
 
 output "runner_agent_role_name" {
   description = "Name of the role used for the ec2 instance for the GitLab runner agent."
-  value       = aws_iam_role.instance.name
+  value       = local.aws_iam_role_instance_name
 }
 
 output "runner_role_arn" {
@@ -51,4 +51,9 @@ output "runner_eip" {
 output "runner_launch_template_name" {
   description = "The name of the runner's launch template."
   value       = aws_launch_template.gitlab_runner_instance.name
+}
+
+output "runner_user_data" {
+  description = "The user data of the Gitlab Runner Agent's launch template."
+  value       = local.template_user_data
 }